refactor(bitcoin-core-sv2): localize v31 JDP handlers and monitor
Move v31.x JDP declare handler and background monitor from shared v32x_v31x sources into v31x-local modules. The only cross-version JDP artifact remaining between v31.x and v32.x is the shared error definition included via #[path].

Remove the now-unused v32x_v31x/job_declaration_protocol/handlers.rs and drop the v32x_v31x namespace marker module.
