+++ bitcoin-core-sv2/src/unix_capnp/v31x/job_declaration_protocol/error.rs
@@ -1,11 +1,11 @@
-//! Error types shared by Bitcoin Core v31.x and v32.x Sv2 Job Declaration Protocol runtimes.
+//! Error types for Bitcoin Core v31.x Sv2 Job Declaration Protocol via capnp over UNIX socket.
 
 use std::path::PathBuf;
 use stratum_core::bitcoin::consensus;
 
 use super::bitcoin_capnp_types::capnp;
 
-/// Errors from the Sv2 Job Declaration Protocol IPC runtime layer.
+/// Errors from the [`crate::unix_capnp::v31x::job_declaration_protocol::BitcoinCoreSv2JDP`] layer.
 #[derive(Debug)]
 pub enum BitcoinCoreSv2JDPError {
     /// Cap'n Proto RPC error.
